feat(pkcs11): generalized PKCS#11 module + card plugin hardening
Extracts the cardedge-specific PKCS#11 module into a standalone librescrs-pkcs11
so every card plugin (rs-eid, pkcs15, cardedge, piv, eu-vrc, rs-health,
opensc, emrtd) exposes keys through one uniform surface.

PKCS#11 module (lib/pkcs11):
- Per-slot std::mutex replaces the global one; sessionMutex held across the
  full body of findObjectsInit/findObjects/findObjectsFinal and
  signInit/sign/signUpdate/signFinal so concurrent closeSession() can't
  invalidate an in-flight session pointer (consistent session→slot order).
- findObjects paginates in 64-slot batches per §5.7.7 with a defensive 4096
  cap against misbehaving tokens.
- signInit enforces cross-slot isolation (key handle from slot A rejected
  on session bound to slot B) and validates PSS mechanism parameters.
- EC signing routed via CKM_ECDSA; raw r||s wrapping handled at the
  provider boundary (see native signing commit).
- Compatibility: reports v2.40 in C_GetFunctionList for SunPKCS11 consumers.

Card plugins:
- SecureBuffer PIN handling throughout; PinStringScrubber RAII ensures
  OPENSSL_cleanse runs on every exit path (normal, throw, early-return).
- PKCS#15 PACE-before-applet-select for contactless eID, CKA_ID / cert
  linking, EC params + point population from EC certificates.
- PIV: AID probe with re-select on reconnect, key-ref/algId derivation
  from cert key type, RSA+ECDSA signData.
- CardEdge: re-verify PIN after SELECT AID resets security status.
- Smartcard: 61xx GET-RESPONSE loop capped at 256 iters / 64 KiB so a
  hostile card can't grow the buffer without bound; SecureBuffer
  exports a canonical PinStringScrubber helper used by all callers.
- Card plugin registry + metadata plumbing for the multi-path trust store
  and widget registry.

+++ lib/smartcard/include/smartcard/apdu.h
@@ -13,7 +13,7 @@ struct APDUCommand
 {
     uint8_t cla, ins, p1, p2;
     std::vector<uint8_t> data; // Lc data (empty = no data field)
-    uint8_t le = 0;            // Expected response length (0 = 256)
+    uint16_t le = 0;           // Expected response length (0 = max, >255 = extended)
     bool hasLe = true;
 
     std::vector<uint8_t> toBytes() const;
